+++ stfblender/stf_resources/stf/stf_instance_mesh/stf_instance_mesh_data.py
@@ -2,13 +2,9 @@ import bpy
 
 
 class STF_Instance_Mesh_Blendshape_Value(bpy.types.PropertyGroup):
-	name: bpy.props.StringProperty(name="Name", options=set())
-	override: bpy.props.BoolProperty(name="Override", default=False, options=set())
-	value: bpy.props.FloatProperty(name="Value", default=0, soft_min=0, soft_max=1, precision=3, subtype="FACTOR")
-	index_on_mesh: bpy.props.IntProperty(name="Index")
+	value: bpy.props.FloatProperty(name="Value", default=0, soft_min=0, soft_max=1, precision=3, subtype="FACTOR", description="Instance specific value for this shape key. To animate this, please animate the Meshes actual shape keys and set this Object as the target with the SlotLink extension.", options=set())
 
 class STF_Instance_Mesh(bpy.types.PropertyGroup):
-	override_blendshape_values: bpy.props.BoolProperty(name="Override Blendshape Values", default=False, options=set())
-	blendshape_values: bpy.props.CollectionProperty(type=STF_Instance_Mesh_Blendshape_Value, name="Blendshape Values", options=set())
+	blendshape_values: bpy.props.CollectionProperty(type=STF_Instance_Mesh_Blendshape_Value, name="Shape Key Values", options=set())
 	active_blendshape: bpy.props.IntProperty(options=set())